Copy language definitions from the tokei project
These have been simplified from the upstream JSON file to remove
features that taxonate will not need. I've also started a CUE file to
generate languages.json more easily and validate contraints on the data
(unimplented).

I plan on adding more data to this structure and changing things a bit,
but the upstream file is licensed under MIT or Apache 2.0:

    Copyright (c) 2016 Aaron Power

To generate the languages file, you can export the CUE data and then I
filter it through Prettier for consistency in formatting:

    $ cue export languages.cue | prettier --parser json > languages.json

See:
- https://github.com/XAMPPRocky/tokei
- https://github.com/XAMPPRocky/tokei/blob/fcf679eab12abbf2463fb57b9e08d31fe4196439/languages.json
- https://cuelang.org/
- https://prettier.io/