Add schema with language field constraints
This requires a name field and will default to the map key if it's not
overridden. I wasn't able to figure out how to require at least one of
the file format markers, so for now, they're all marked as optional. If
a marker *is* specified for a language, one or more values is required.