Add constraint to require at least one marker
The `marker_is_defined` constraint unfortunately does not work
at the moment. It does work in the CUE Playground which is running
v0.0.16-0.20200323154021-aa5e0c739758 and Marcel van Lohuizen indicated
that a rewrite of the evaluator is underway, so I'll keep an eye on this
one and see if I can get it going again in the future.

At some point, a proper topoplogical sort will also be introduced to
better guarantee the sort order. Right now, tokens that match builtins
from the standard library will jump to the top of the map keys under
certain scenarios.