Set up command line argument handling and logging
I'm starting to flesh out the desired user interface for the CLI
application and want to have logging facilities in place. I've
configured clap to allow for controlling both the debug log level
and color settings (which I'll be able to re-use in my own output).
Environment variables will take precedence, when specified.