Implement collecting paths into a HashSet
This properly implements the behavior I was expecting where:

- the list of paths will not contain duplicate entries

- it will append entries from STDIN if explicitly requested (I couldn't
  figure out the implicit logic to work in an intuitive way)

- it will default to just the current working directory if no other
  paths have been specified

I believe that paths are not guaranteed to be valid UTF-8, so the proper
way to handle that is to use OsString values instead of String values.
Note that I've also dropped the external bstr crate dependency, since
I'm not 100% confident on how to convert between the two, or if it will
even be beneficial/necessary. I can revist that in the future once the
real parsing functionality is implemented.

See:
- https://github.com/clap-rs/clap/issues/751