Use to_owned instead of to_string for semantic clarity
There's no performance difference here, as they're treated exactly the
same, but I feel like there is a bit of a sematic difference which makes
the intent more clear. The fact that I do have the `String::from` uses
in here might make it a bit muddier, but that's better than adding a
closure (which clippy will complain about).

See:
- https://users.rust-lang.org/t/what-is-the-idiomatic-way-to-convert-str-to-string/12160/6
- https://users.rust-lang.org/t/to-string-vs-to-owned-for-string-literals/1441/7