Document how language support is configured
Typing this out makes me realize that, yes, it's a convoluted process.
It mostly came from my desire to also explore and learn more about CUE,
and gain experience with Serde as well. If I end up struggling too much,
I may just encode these definitions in Rust proper. Realistically, I'll
end up going down that road anyway, as most of the constraints I've put
into place with CUE can be replicated easily with testing, and there's
no configuration sharing to be concerned about, so it's not necessarily
the best use case for the language. It's not a complex data structure.

Regardless, I'll push this path forward a bit more, and will be glad to
have the commits stored away for posterity.