Update the clippy-check name for clarity
The summary output will always create a new check, so let's use an
explicit name to make it more clear what you'll see when clicking on it.
We'll soon find out what the pass/fail markers look like for both "Lint"
and the "Clippy Results" check when there's a legit lint suggestion.