Pass the HashSet directly when building a Config
For now, it's a public implementation detail, and we want the behavior
of a HashSet when figuring out the default behavior within the clap app,
so might as well pass it directly rather than use a slice of Strings.
Otherwise, we'd have to do some conversion like this:

   let path_values: Vec<String> = paths.drain().collect();

...which seems unnecessary at the moment since it would be converted
right back.