Address clippy lints
I'm cheating a bit on the errors doc lint, but plan on going through
to add all sorts of docs once I've settled on a structure for all of
these pieces. The lifetime annotations were essentially copy/pasted from
the upstream builder pattern example, so hopefully relying on lifetime
elision is okay in this case...it is much cleaner.

See:
- https://rust-lang.github.io/rust-clippy/master/index.html#missing_errors_doc
- https://rust-lang.github.io/rust-clippy/master/index.html#needless_lifetimes
- https://rust-lang.github.io/rust-clippy/master/index.html#must_use_candidate