Add flag to display file names without language
This will be useful for the CI use case where I'd like to identify files
to send on to a formatter/linter as part of automated testing.

Again, the nested conditional here isn't great, since that won't scale
if other flags are added. Need to figure out a clean pattern for these
feature toggles.